Test `overwrite` flag for state IO
