+++ modelrunner/tests/test_trajectory.py
@@ -5,46 +5,28 @@
 import numpy as np
 import pytest
 
-from modelrunner.state import (
-    ArrayCollectionState,
-    ArrayState,
-    DictState,
-    ObjectState,
-    StateBase,
-)
+from modelrunner.trajectory import Trajectory, TrajectoryWriter
 
-EXTENSIONS = ["json", "yaml", "zarr"]
-
-
-def get_states():
-    """generate multiple states"""
-    a = np.arange(5)
-    b = np.random.random(size=3)
-    o = {"list": [1, 2], "bool": True}
-    return [
-        ObjectState(o),
-        ArrayState(a),
-        ArrayCollectionState((a, b), labels=["a", "b"]),
-        DictState({"o": ObjectState(o), "a": ArrayState(a)}),
-    ]
+from .test_state import EXTENSIONS, get_states
 
 
 @pytest.mark.parametrize("state", get_states())
 @pytest.mark.parametrize("ext", EXTENSIONS)
-def test_state_io(state, ext, tmp_path):
-    """test simple state IO"""
+def test_trajectory(state, ext, tmp_path):
+    """test simple trajecotry writing"""
     path = tmp_path / ("file." + ext)
-    state.to_file(path)
-    state2 = StateBase.from_file(path)
-    assert state == state2
 
+    with TrajectoryWriter(path, attrs={"test": "yes"}) as write:
+        write(state, 1)
+        write(state)
 
-@pytest.mark.parametrize("state_cls", [DictState, ObjectState, ArrayCollectionState])
-@pytest.mark.parametrize("ext", EXTENSIONS)
-def test_empty_state_io(state_cls, ext, tmp_path):
-    """test simple state IO"""
-    state = state_cls()
-    path = tmp_path / ("file." + ext)
-    state.to_file(path)
-    state2 = StateBase.from_file(path)
-    assert state == state2
+    for ret_copy in [True, False]:
+        traj = Trajectory(path, ret_copy=ret_copy)
+        assert len(traj) == 2
+        np.testing.assert_allclose(traj.times, [1, 2])
+        assert traj[1] == state
+        assert traj[-1] == state
+        assert traj.attributes == {"test": "yes"}
+
+        for s in traj:
+            assert s == state
